# fota-application/src/main/resources/static/swagger-coding-challenge-fota.yaml
swagger: "2.0"
info:
  description: "API to query installable and incompatible features on trucks."
  version: "1.0.0"
  title: "MAN FOTA Challenge"
host: "fota.com"
tags:
- name: "fota"
  description: "Query possible features per vehicle and possible vehicles per feature"
paths:
  /fota/vehicles/{vin}/installable:
    get:
      tags:
      - "fota"
      summary: "Find installable features by VIN"
      description: "gives all the features that can be installed for the corresponding vin"
      operationId: "getInstallable"
      produces:
      - "application/json"
      parameters:
      - name: "vin"
        in: "path"
        description: "Vehicle identification number"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO
        404:
          description: "Not found"

  /fota/vehicles/{vin}/incompatible:
    get:
      tags:
      - "fota"
      summary: "Find incompatible features by VIN"
      description: "gives all the features that cannot be installed for the corresponding vin"
      operationId: "getIncompatible"
      produces:
      - "application/json"
      parameters:
      - name: "vin"
        in: "path"
        description: "Vehicle identification number"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO
        404:
          description: "Not found"

  /fota/vehicles/{vin}:
    get:
      tags:
      - "fota"
      summary: "Find all features by VIN"
      description: "gives all features that can/cannot be installed for the corresponding vin"
      operationId: "getFeatures"
      produces:
      - "application/json"
      parameters:
      - name: "vin"
        in: "path"
        description: "Vehicle identification number"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO
        404:
          description: "Not found"

  /fota/vehicles:
    get:
      tags:
      - "fota"
      summary: "Find all vehicles"
      description: "returns a list of all vehicles"
      operationId: "getVehicles"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO

  /fota/features/{feature}/installable:
    get:
      tags:
      - "fota"
      summary: "Find installable VIN's by feature code"
      description: "gives all the vins that can install the corresponding feature"
      operationId: "getVinInstallable"
      produces:
      - "application/json"
      parameters:
      - name: "feature"
        in: "path"
        description: "Correspondent feature code"
        required: true
        type: "string"
        format: "string"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO
        404:
          description: "Not found"

  /fota/features/{feature}/incompatible:
    get:
      tags:
      - "fota"
      summary: "Find incompatible VIN's by feature code"
      description: "gives all the vins that cannot install the corresponding feature"
      operationId: "getVinIncompatible"
      produces:
      - "application/json"
      parameters:
      - name: "feature"
        in: "path"
        description: "Correspondent feature code"
        required: true
        type: "string"
        format: "string"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO
        404:
          description: "Not found"

  /fota/features/{feature}:
    get:
      tags:
      - "fota"
      summary: "Find all VIN's by feature code"
      description: "gives all the vins that can install the corresponding feature"
      operationId: "getAllVin"
      produces:
      - "application/json"
      parameters:
      - name: "feature"
        in: "path"
        description: "Correspondent feature code"
        required: true
        type: "string"
        format: "string"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO
        404:
          description: "Not found"

  /fota/features:
    get:
      tags:
      - "fota"
      summary: "Find all features"
      description: "returns a list of all feature codes"
      operationId: "getAllFeatures"
      produces:
      - "application/json"
      parameters:
      - name: "feature"
        in: "path"
        description: "Correspondent feature code"
        required: true
        type: "string"
        format: "string"
      responses:
        200:
          description: "successful operation"
          #schema:
            #TODO

          